requirements.txt
tensorflow==2.20.0
streamlit==1.53.1
fonttools==4.61.1
Pillow==12.1.0
pytest==9.0.2
